refactor: single source of truth for AI model registry (src/shared/aiModels)
Introduce runtime-neutral src/shared/ (importable by both Vite frontend via
@/shared and the Bun API via ../../src/shared/*.js). Move the AI_MODELS registry
+ derived types there; src/types/aiModels re-exports it, and api/_utils/_aiModels
imports it while keeping server-only provider wiring (getModelInstance, etc.).
Removes the explicit 'duplicated from src/types/aiModels.ts' copy.

Verified: frontend build + 42 AI API tests (incl. model validation + provider options).
