+++ src/types/aiModels.ts
@@ -1,45 +1,13 @@
-// Shared AI model types and constants
-
-// Single source of truth for AI models
-export const AI_MODELS = {
-  "sonnet-4.6": {
-    name: "sonnet-4.6",
-    provider: "Anthropic",
-  },
-  "gpt-5.5": {
-    name: "gpt-5.5",
-    provider: "OpenAI",
-  },
-  "gemini-3-flash": {
-    name: "gemini-3-flash",
-    provider: "Google",
-  },
-  "gemini-3.1-pro-preview": {
-    name: "gemini-3.1-pro-preview",
-    provider: "Google",
-  },
-} as const;
-
-// Derived types
-export type AIModel = keyof typeof AI_MODELS | null;
-export type SupportedModel = keyof typeof AI_MODELS;
-
-// Derived arrays
-export const SUPPORTED_AI_MODELS = Object.keys(AI_MODELS) as SupportedModel[];
-
-// Model metadata for UI display
-export interface AIModelInfo {
-  id: SupportedModel;
-  name: string;
-  provider: string;
-}
-
-export const AI_MODEL_METADATA: AIModelInfo[] = Object.entries(AI_MODELS).map(
-  ([id, info]) => ({
-    id: id as SupportedModel,
-    ...info,
-  })
-);
-
-// Default model
-export const DEFAULT_AI_MODEL: SupportedModel = "gpt-5.5";
+// Re-export the runtime-neutral AI model registry from the shared module.
+// Kept as a stable import path (`@/types/aiModels`) for existing frontend code.
+export {
+  AI_MODELS,
+  SUPPORTED_AI_MODELS,
+  AI_MODEL_METADATA,
+  DEFAULT_AI_MODEL,
+} from "@/shared/aiModels";
+export type {
+  AIModel,
+  SupportedModel,
+  AIModelInfo,
+} from "@/shared/aiModels";
